make zmugjson work with simplejson 1.3 *and* 1.7. In 1.7, the object_hook is called everytime a new "object" or dict is found.
Take for instance the following string: 
‘{”foobar”:”baz”,”comp”:{”key”:10},”array”:[10,20,30]}’

1.3 would pass the above string into the object_hook as is. With 1.7 the
following happens:
first {u’key’: 10} is passed in, then {u’foobar’: u’baz’, u’array’: [10, 20,
30], u’comp’: None} is passed in.

In order to fix this, I continue to look for the key "Categories", but if
the argument does not contain it, I return the string unaltered. Ultimately,
I get the same result.